refactor: move code from services to root
This is the first of a couple refactors to
abstract as much service code into a separate
package in order to reduce the amount of
code duplication.

This code was essentially copy pasted into
all of the services because we didn't have
a good way to abstract it until now.