fix: migration script for post tags
The only way that post tags get added to our database
is if the post gets uploaded.  If there is a post in our
system that already has tags, the tags will show up in the post
but not in the blog tag filter.

This script backports tags for all posts that already had tags
in their system.