feat: dates migration and reorg one-off scripts
Since we've been needing to perform more complex migrations that depend
on golang, I thought it would be a good idea to have a "home" for our
one-off scripts.

I like the idea of keeping them in git so we can reference them for
future scripts or just for historical reasons.