refactor(imgs): redesign blog page for imgs
After chatting with the pico members, it became clear that the photo
album direction I was heading in was not going to super useful.  Instead
we decided to revert the design back to how the other pico services
function.  So now it's just a stream of images not grouped by tags.