bug(upload): removed user field from shared file uploader struct
When a user attempts to upload files to our server, the
logic we use to perform file uploads is baked into a few similar
structs (`ScpUploadHandler` and `UploadImgHandler`).  In those
structs we authenticate their `ssh.Session` by using their name
to find a user record in our db.

We mistakenly had a user field on those shared structs which would
sometimes result in a race condition where user A would upload files to
user B's blog.

We have removed the user field and instead leverage
`ssh.Session.Context` to set and access the user object for our file
uploaders.