refactor(pastes): add col `expires_at` to posts
Currently we figure out how to delete pastes by having a cron goroutine
check on an interval whether or not posts are older than 3 days.
Instead, we want to provide the ability for that delete time to be
determined by individual records.

So we added `expires_at` column to our `posts` table.

I also extended the default expiration from 3 to 7 days.