feat: cannot link to a project that is also a link
A project linked to another project which is also linked to a
project is forbidden.  CI/CD Example:
        - ProjectProd links to ProjectStaging
        - ProjectStaging links to ProjectMain
        - We merge `main` and trigger a deploy which uploads to ProjectMain
        - All three get updated immediately
This scenario was not the intent of our CI/CD.  What we actually
wanted was to create a snapshot of ProjectMain and have ProjectStaging
link to the snapshot, but that's not the intended design of pgs.

So we want to close that gap here.