mdparser: implicitly capture hashtags in document
When they're found, they will be rendered as so:
`<span class="hashtag">...</span>`.

They are ignored in the following:
* block code sections
* block quote sections
* span code sections

This list can be expanded.

We then collect all of the matching hashtags and provide them as
metadata, but only if there is no explicit frontmatter for them.
So a workaround for undesirable hashtag detection is to use front-matter
to specify `tags: []` or similar.