feat(auth): subscribe to pico's metric-drain pipe
We have a lot of services that need to record site usage analytics so we
need a distributed way to receive these events.

As a result, all of our site visits will be sent through our pipe
service.  Then inside of our auth service we process the visits and
store them in our analytics table.

This overloads the auth service since it's now serving as a destination
for our metric-drain, but we think that's okay for now.