feat(feeds): cli
A remote CLI with some handy options: ls, rm, and run

`run {filename}` is the most interesting because it will run the feed
digest post immediately, ignoring last digest time validation.

This will be useful for people to debug their posts since that is a
pain point for a lot of users.