fix(tui.analytics): allow users to enable analytics
The core issue here is around focus management.  When a user enters the
analytics page, we fetch site stats regardless if they have analytics
enabled which then applies focus to the left page.  However for users
witohut analytics enabled we do not focus the left pane which means all
key events are lost.  In vaxis if we try to focus a widget that doesn't
exist we lose the key events in our page widget.