fix(pgs): ability to create multiple projects in single ssh session
Previously we were caching the project when a user uploaded a bunch of
files.  This was an optimization that didn't account for when a user has
an active sftp connection where they create multiple projects in a
single session.  This code change performs an extra check to ensure we
are always creating new projects when necessary.